requirements.txt
mss
pillow
pystray
